feito

+++ routes/delete.php
@@ -1,7 +1,17 @@
 <?php 
-require_once "../db.php";
+require_once "../config.php";
+$db = new Database();
 
-$id = $POST['id'];
-$query = "DELETE FROM notas WHERE id = '$id'";
+if (assert_key_exists(['id'], $_POST)) {
+    $query = "DELETE FROM notas WHERE id = '$id'";
 
-$conn->query($query);
+    $db->query($query);
+} else {
+    $db->close();    
+    header("500 Internal Server Error");
+    die();
+}
+
+$notas = $db->get_many('notas');
+echo json_encode($notas);
+$db->close();
